Fix docker build (#138)
* Fix docker build

The command 'pnpm install' would fail, because it also executed
the lifecycle script 'pnpm build' automatically. But judging from the
rest of the Dockerfile, this is not intended. 'pnpm build' is executed
later in the Dockerfile explicitly, when the necessary preparations
have been made.

* update pnpm-lock.yaml

This is necessary because the lockfile format has had a major version
bumb from 5 to 6. Trying to install with --frozen-lockfile results
in an error because of that.